Move update models to a dedicated subfolder
