+++ src/langium-model-server/source/model/action/rename.ts
@@ -1,4 +1,4 @@
-import type { Update } from '../update'
+import type { Update } from '../update/update'
 import type * as id from '../../../semantic/identity'
 
 export type Rename = Update<id.NamedSemanticIdentity, 'RENAMED'>
