+++ src/langium-model-server/source/model/index.ts
@@ -2,7 +2,7 @@ export * from './action'
 export * from './api/api-response'
 export * from './api/highlight-response'
 export * from './api/semantic-id-response'
-export * from './array-update'
-export * from './element-update'
-export * from './update'
+export * from './update/array-update'
+export * from './update/element-update'
+export * from './update/update'
